[scheduler][styled] Move the private components to `src/internals/components` (#18894)
