+++ packages/x-scheduler/src/material/event-calendar/EventCalendar.tsx
@@ -5,14 +5,14 @@ import { EventCalendarProps } from './EventCalendar.types';
 import { WeekView } from '../week-view/WeekView';
 import { AgendaView } from '../agenda-view';
 import { DayView } from '../day-view/DayView';
-import { HeaderToolbar } from '../header-toolbar';
 import { TranslationsProvider } from '../internals/utils/TranslationsContext';
 import { useSelector } from '../../base-ui-copy/utils/store';
 import { selectors } from './store';
 import { EventCalendarContext } from '../internals/hooks/useEventCalendarContext';
 import { MonthView } from '../month-view';
-import { DateNavigator } from '../date-navigator/DateNavigator';
-import { ResourceLegend } from '../internals/components/resource-legend/ResourceLegend';
+import { HeaderToolbar } from '../internals/components/header-toolbar';
+import { DateNavigator } from '../internals/components/date-navigator';
+import { ResourceLegend } from '../internals/components/resource-legend';
 import { useEventCalendar } from './useEventCalendar';
 import '../index.css';
 import './EventCalendar.css';
