+++ packages/x-scheduler/src/material/internals/components/date-navigator/DateNavigator.css
