+++ packages/x-scheduler/src/material/internals/components/header-toolbar/HeaderToolbar.css
@@ -1,4 +1,4 @@
-@import '../styles/tokens.css';
+@import '../../../styles/tokens.css';
 
 .HeaderToolbarContainer {
   display: flex;
