+++ packages/x-scheduler/src/material/internals/components/header-toolbar/HeaderToolbar.tsx
@@ -3,8 +3,8 @@ import * as React from 'react';
 import clsx from 'clsx';
 import { HeaderToolbarProps } from './HeaderToolbar.types';
 import { ViewSwitcher } from './view-switcher';
-import { useTranslations } from '../internals/utils/TranslationsContext';
-import { useEventCalendarContext } from '../internals/hooks/useEventCalendarContext';
+import { useTranslations } from '../../utils/TranslationsContext';
+import { useEventCalendarContext } from '../../hooks/useEventCalendarContext';
 import './HeaderToolbar.css';
 
 export const HeaderToolbar = React.forwardRef(function HeaderToolbar(
