+++ packages/x-scheduler/src/material/internals/components/header-toolbar/HeaderToolbar.types.ts
