+++ packages/x-scheduler/src/material/internals/components/header-toolbar/index.ts
