+++ packages/x-scheduler/src/material/internals/components/header-toolbar/view-switcher/ViewSwitcher.css
@@ -1,4 +1,4 @@
-@import '../../styles/tokens.css';
+@import '../../../../styles/tokens.css';
 
 .ViewSwitcherMenuBar {
   display: flex;
