+++ packages/x-scheduler/src/material/internals/components/header-toolbar/view-switcher/index.ts
